requirements.txt
streamlit==1.41.1
yfinance==1.2.0
pyyaml==6.0.2
pandas==2.2.3
openpyxl==3.1.5
requests==2.32.3
xlrd==2.0.1
